Interfaces fix, typo fix, renaming
