+++ CloudLayout/Interfaces/IInputOptions.cs
@@ -1,0 +1,11 @@
+﻿using System.Drawing;
+
+namespace CloudLayout.Interfaces
+{
+    public interface IInputOptions
+    {
+        Point CenterPoint { get; set; }
+
+        int Size { get; set; }
+    }
+}
