+++ CloudLayout/Interfaces/ILayout.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace CloudLayout.Interfaces
+{
+    public interface ILayout
+    {
+        bool PutNextRectangle(SizeF size, out RectangleF rectangle);
+    }
+}
