+++ CloudLayout/Interfaces/ISpiralDrawer.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace CloudLayout.Interfaces
+{
+    public interface ISpiralDrawer
+    {
+        List<PointF> GetSpiralPoints(Point center);
+    }
+}
