+++ TagsCloudContainer/CustomOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace TagsCloudContainer
+﻿using TagsCloudContainer.Interfaces;
+
+namespace TagsCloudContainer
 {
     public class CustomOptions : ICustomOptions
     {
@@ -11,18 +13,5 @@
         public string BackgroundColor { get; set; } = string.Empty;
         public string FontColor { get; set; } = string.Empty;
         public string ExcludedParticals { get; set; } = string.Empty;
-    }
-
-    public interface ICustomOptions
-    {
-        int PictureSize { get; set; }
-        int FontSize { get; set; }
-        string TextsPath { get; set; }
-        string WordsFileName { get; set; }
-        string BoringWordsName { get; set; }
-        string Font { get; set; }
-        string BackgroundColor { get; set; }
-        string FontColor { get; set; }
-        string ExcludedParticals { get; set; }
     }
 }
