+++ TagsCloudContainer/Interfaces/IConverter.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloudContainer.Interfaces
+{
+    public interface IConverter
+    {
+        Dictionary<string, int> GetWordsInFile(ICustomOptions options);
+    }
+}
