+++ TagsCloudContainer/Interfaces/IWordSizeCalculator.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudContainer.Interfaces
+{
+    public interface IWordSizeCalculator
+    {
+        public Dictionary<string, Font> CalculateSize(Dictionary<string, int> input, ICustomOptions options);
+    }
+}
